Group functions working properly
Group functions are working properly now. user can leave a group and
join a group. Now the it needs to check to see if the user is already
in a group. If not it makes a user join a group.  Added some stuff to
the group funds request page. added some questions about what the
project is . still needs to be edited